+++ CHANGELOG.md
@@ -232,3 +232,4 @@
 - 2023-12-30: Test main, descend, sender, ready pool, +3
 - 2023-12-31: Test deal, daemon, README, bettercontact, +3
 - 2024-01-01: Docs sender, conf, users, find email, +3
+- 2024-01-02: Bench contacts, find email, leads, lead, +4